refactor: use aspida for API
