+++ src/features/forms/components/EachForm/hooks.ts
@@ -1,10 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
 import { getForm } from '../../services/getForm';
-import { FormId } from '../../types';
 
-export const useForm = (id: FormId) => {
-  const { data } = useQuery([id, 'form'], () => getForm(id));
-
-  return data;
-};
+export const useForm = (id: number) =>
+  useQuery([id, 'form'], () => getForm(id)).data;
