+++ src/features/forms/components/FormInfoList/presenter.tsx
@@ -3,7 +3,7 @@ import { SimpleGrid, Text } from '@chakra-ui/react';
 import { Alert } from '@/components/Alert';
 import { LikeLink } from '@/components/LikeLink';
 
-import type { FormInfo } from '../../types';
+import type { FormInfo } from '../../types/formInfo';
 
 type Props = {
   forms?: FormInfo[];
@@ -17,7 +17,7 @@ export const Presenter = ({ forms = [] }: Props) =>
       {forms.map((form) => (
         <LikeLink
           key={form.id}
-          title={form.formName}
+          title={form.form_name}
           path={`/forms/${form.id}`}
         >
           <Text fontSize="md" noOfLines={3}>
