--- src/features/forms/types/form.ts
@@ -1,23 +1,0 @@
-import { isNonEmptyString, NonEmptyString } from '@/types';
-import { isString } from '@/types/assertString';
-
-import { FormId, isFormId } from './formId';
-import { isQuestionList, Question } from './question';
-
-export type Form = {
-  id: FormId;
-  formName: NonEmptyString;
-  description?: string;
-  questions: Question[];
-};
-
-export const isForm = (arg: unknown): arg is Form => {
-  const f = arg as Form;
-
-  return (
-    isFormId(f?.id) &&
-    isNonEmptyString(f?.formName) &&
-    isString(f?.description) &&
-    isQuestionList(f?.questions ?? [])
-  );
-};
