--- src/features/forms/types/index.ts
@@ -1,5 +1,0 @@
-export * from './form';
-export * from './formId';
-export * from './question';
-export * from './questionId';
-export * from './formInfo';
